float

+++ app1/src/store_price.py
@@ -53,7 +53,7 @@ def sum_of_prices(product_id):
     print(str(products)[:500])
     product_id = str(product_id)
     items = products[product_id]
-    return round(sum(int(item[2]) for item in items))
+    return round(sum(float(item[2]) for item in items))
 
 
 """
